requirements.txt
Django==1.10.5
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
djangorestframework==3.5.4
django-filter==1.0.1
Pillow==3.1.0
wheel==0.24.0
slackclient==1.0.5
django_tables2
pandas
